+++ src/Var.jl
@@ -14,6 +14,7 @@ import ..Utils:
     seconds_to_prettystr,
     squeeze,
     split_by_season,
+    split_by_season_across_time,
     time_to_date,
     date_to_time,
     _data_at_dim_vals,
@@ -48,6 +49,7 @@ export OutputVar,
     integrate_lat,
     isempty,
     split_by_season,
+    split_by_season_across_time,
     bias,
     global_bias,
     squared_error,
@@ -1288,50 +1290,119 @@ end
 
 Return a vector of four `OutputVar`s split by season.
 
-The months of the seasons are March to May, June to August, September to November, and
-December to February. The order of the vector is MAM, JJA, SON, and DJF. If there are no
-dates found for a season, then the `OutputVar` for that season will be an empty `OutputVar`.
+The months of the seasons are March to May (MAM), June to August (JJA), September to
+November (SON), and December to February (JDF). The order of the vector is MAM, JJA, SON,
+and DJF. If there are no dates found for a season, then the `OutputVar` for that season will
+be an empty `OutputVar`.
 
 The function will use the start date in `var.attributes["start_date"]`. The unit of time is
-expected to be second. Also, the interpolations will be inaccurate in time intervals
-outside of their respective season for the returned `OutputVar`s.
+expected to be second.
+
+!!! note "Interpolating between seasons"
+    Interpolations will be inaccurate in time intervals outside of their respective season
+    for the returned `OutputVar`s. For example, if an `OutputVar` has the dates 2010-2-1,
+    2010-3-1, 2010-4-1, and 2011-2-1 after splitting by seasons, then any interpolation in
+    time between the dates 2010-4-1 and 2011-2-1 will be inaccurate.
+
+This function differs from `split_by_season_across_time` as `split_by_season_across_time`
+splits dates by season for each year.
 """
 function split_by_season(var::OutputVar)
-    # Check time exists and unit is second
-    has_time(var) || error("Time is not a dimension in var")
-    dim_units(var, time_name(var)) == "s" ||
-        error("Unit for time is not second")
-
-    # Check start date exists
-    haskey(var.attributes, "start_date") ?
-    start_date = Dates.DateTime(var.attributes["start_date"]) :
-    error("Start date is not found in var")
+    _check_time_dim(var::OutputVar)
+    start_date = Dates.DateTime(var.attributes["start_date"])
 
     season_dates = split_by_season(time_to_date.(start_date, times(var)))
     season_times =
         (date_to_time.(start_date, season) for season in season_dates)
 
-    # Split data according to seasons
-    season_data = (
+    return _split_along_dim(var, time_name(var), season_times)
+end
+
+"""
+    split_by_season_across_time(var::OutputVar)
+
+Split `var` into `OutputVar`s representing seasons, sorted in chronological order. Each
+`OutputVar` corresponds to a single season, and the ordering of the `OutputVar`s is
+determined by the dates of the season. The return type is a vector of `OutputVar`s.
+
+The months of the seasons are March to May (MAM), June to August (JJA),
+September to November (SON), and December to February (DJF). If there are no
+dates found for a season, then the `OutputVar` for that season will be an empty
+`OutputVar`. The first `OutputVar` is guaranteed to not be empty.
+
+The function will use the start date in `var.attributes["start_date"]`. The unit of time is
+expected to be second.
+
+This function differs from `split_by_season` as `split_by_season` splits dates by
+season and ignores that seasons can come from different years.
+"""
+function split_by_season_across_time(var::OutputVar)
+    _check_time_dim(var::OutputVar)
+    start_date = Dates.DateTime(var.attributes["start_date"])
+
+    seasons_across_year_dates =
+        split_by_season_across_time(time_to_date.(start_date, times(var)))
+    seasons_across_year_times = (
+        date_to_time.(start_date, season) for
+        season in seasons_across_year_dates
+    )
+
+    return _split_along_dim(var, time_name(var), seasons_across_year_times)
+end
+
+"""
+    check_time_dim(var::OutputVar)
+
+Check time dimension exists, unit for the time dimension is second, and a
+start date is present.
+"""
+function _check_time_dim(var::OutputVar)
+    has_time(var) || error("Time is not a dimension in var")
+    dim_units(var, time_name(var)) == "s" ||
+        error("Unit for time is not second")
+    haskey(var.attributes, "start_date") ||
+        error("Start date is not found in var")
+    return nothing
+end
+
+"""
+    _split_along_dim(var::OutputVar, dim_name, split_vectors)
+
+Given `dim_name` in `var`, split the `OutputVar` by the values in `split_vectors`
+and return a vector of `OutputVar`s.
+
+For example, if `dim_name = "time" and `split_vectors = [[0.0, 3.0], [2.0,
+4.0]]`, the result is a vector of two `OutputVar`s, where the first OutputVar
+has a time dimension of `[0.0, 3.0]` and the second OutputVar has a time
+dimension of `[2.0, 4.0]`.
+
+If the vector in `split_vectors` is empty, then an empty OutputVar is returned.
+Additonally, there is no checks that are performed in the values in the vectors
+in `split_vectors` as the nearest values in `var.dims[dim_name]` are used for
+splitting.
+"""
+function _split_along_dim(var::OutputVar, dim_name, split_vectors)
+    # Split data by vectors in split_vectors
+    split_data = (
         collect(
             _data_at_dim_vals(
                 var.data,
-                times(var),
-                var.dim2index[time_name(var)],
-                season_time,
+                var.dims[dim_name],
+                var.dim2index[dim_name],
+                split,
             ),
-        ) for season_time in season_times
+        ) for split in split_vectors
     )
 
     # Construct an OutputVar for each season
-    return map(season_times, season_data) do time, data
-        if isempty(time)
+    return map(split_vectors, split_data) do split, data
+        if isempty(split)
             dims = empty(var.dims)
             data = similar(var.data, 0)
             return OutputVar(dims, data)
         end
         ret_dims = deepcopy(var.dims)
-        ret_dims[time_name(var)] = time
+        ret_dims[dim_name] = split
         remake(var, dims = ret_dims, data = data)
     end
 end